[IMP] mail: move code to models (MessageComponent)
Task-3000365
