+++ addons/mail/static/src/models/message_list_view.js
@@ -107,8 +107,7 @@ registerModel({
             }
             if (
                 this.threadViewOwner.lastMessageView &&
-                this.threadViewOwner.lastMessageView.component &&
-                this.threadViewOwner.lastMessageView.component.isPartiallyVisible()
+                this.threadViewOwner.lastMessageView.isPartiallyVisible()
             ) {
                 this.threadViewOwner.handleVisibleMessage(
                     this.threadViewOwner.lastMessageView.message,
